refactor: rename HomeError to ComponentError
